+++ packages/ui/src/Detail/index.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { detail } from '../utils/typography';
+
+export const Detail = styled.span`
+  ${detail}
+`;
