+++ packages/ui/src/H1/index.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import { h1 } from '../utils/typography';
+
+/**
+ * Renders a styled <h1 />. Pass the `as` prop to use a different HTML element
+ * with the same styling.
+ */
+export const H1 = styled.h1`
+  ${h1}
+`;
