+++ packages/ui/src/H2/index.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import { h2 } from '../utils/typography';
+
+/**
+ * Renders a styled <h2 />. Pass the `as` prop to use a different HTML element
+ * with the same styling.
+ */
+export const H2 = styled.h2`
+  ${h2}
+`;
