+++ packages/ui/src/H3/index.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import { h3 } from '../utils/typography';
+
+/**
+ * Renders a styled <h3 />. Pass the `as` prop to use a different HTML element
+ * with the same styling.
+ */
+export const H3 = styled.h1`
+  ${h3}
+`;
