+++ packages/ui/src/H4/index.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import { h4 } from '../utils/typography';
+
+/**
+ * Renders a styled <h4 />. Pass the `as` prop to use a different HTML element
+ * with the same styling.
+ */
+export const H4 = styled.h1`
+  ${h4}
+`;
